Separated Capturable check, implemented capture
The capturable check has been separated between checking the whole board and checking one piece. The capture piece code needs to receive a piece to be moved, piece to be captured, and position. The capture check code needs to output these in a way that can be used to select stuff.